requirements.txt
streamlit
pandas
numpy
xgboost
scikit-learn
joblib
google-genai